cluster listener refactor
