+++ nsdb-cluster/src/main/scala/io/radicalbit/nsdb/cluster/NSDbActors.scala
@@ -57,7 +57,6 @@ trait NSDbActors {
 
     DistributedData(system).replicator
 
-    system.actorOf(ClusterListener.props(true),
-                   name = s"cluster-listener_${createNodeName(Cluster(system).selfMember)}")
+    system.actorOf(Props[ClusterListener], name = s"cluster-listener_${createNodeName(Cluster(system).selfMember)}")
   }
 }
